BasitHesapMakinesi düzenlendi.

+++ main.py
@@ -3,15 +3,15 @@ second_number = float(input("İkinci Sayınızı giriniz : "))
 which_procces  = input("toplama(+),çıkarma(-),çarpma(*),bölme(/) :")
 if which_procces == "+":
     def toplama(x,y):
-        print(int(x + y))
+        print(x + y)
     toplama(first_number,second_number)
 elif which_procces == "-":
     def çıkarma(x,y):
-        print(int(x - y))
+        print(x - y)
     çıkarma(first_number,second_number)
 elif which_procces == "*":
     def çarpma(x,y):
-        print(int(x * y))
+        print(x * y)
     çarpma(first_number,second_number)
 elif which_procces == "/":
     def bölme(x,y):
